**High level description:**
This ticket adjusts the logic used by the "All Foreign Countries" option for Place of performance scope in advanced search, preventing US-based award POPs from matching.

**Technical details:**
Uses an OR between pop_country_name and pop_country_code to further filter out those that shouldn't match.

**Link to JIRA Ticket:**
[DEV-1460](https://federal-spending-transparency.atlassian.net/browse/DEV-1460

The following are ALL required for the PR to be merged:
- [ ] Backend review completed
- [NA] Unit & integration tests updated with relevant test cases 

**Performance:**

**/v2/spending_by_award/**

REQUEST:

{
	"fields":["Award ID", "Recipient Name", "Start Date", "End Date", "Award Amount", "Awarding Agency"],
	"filters":{
		"agencies":[{"type": "awarding", "tier": "toptier", "name": "Department of Defense"}],
		"award_amounts": [{"lower_bound": 500000000}],
		"award_type_codes": ["A", "B", "C", "D"],
		"place_of_performance_scope": "foreign",
		"time_period": [{"start_date": "2008-10-01", "end_date": "2009-09-30"}]
	},
	"limit":60,
	"subawards":false
}

Before change: 952 ms
After change: 939 ms


REQUEST:

{
	"fields":["Award ID", "Recipient Name", "Start Date", "End Date", "Award Amount", "Awarding Agency"],
	"filters":{
		"agencies":[{"type": "awarding", "tier": "toptier", "name": "Department of Homeland Security"}],
		"award_amounts": [{"lower_bound": 500001}],
		"award_type_codes": ["A", "B", "C", "D"],
		"place_of_performance_scope": "foreign",
		"time_period": [{"start_date": "2008-10-01", "end_date": "2009-09-30"}]
	},
	"limit":60,
	"subawards":false
}

Before change: 1534 ms
After change: 1548 ms



**/v2/spending_over_time/**
REQUEST:
{
	"fields":["Award ID", "Recipient Name", "Start Date", "End Date", "Award Amount", "Awarding Agency"],
	"group":"quarter",
	"filters":{
		"keyword":"pizza",
		"place_of_performance_scope": "foreign"
	},
	"limit":60,
	"subawards":false
}


Before change: 978 ms
After change: 971 ms


**Subawards**

/v2/spending_by_award/
REQUEST:

{
	"fields":["Sub-Award ID", "Sub-Award Date", "Sub-Award Amount", "Awarding Agency"],
	"filters":{
		"agencies":[{"type": "awarding", "tier": "toptier", "name": "Department of Defense"}],
		"award_type_codes": ["A", "B", "C", "D"],
		"place_of_performance_scope": "foreign",
		"time_period": [{"start_date": "2008-10-01", "end_date": "2009-09-30"},
						{"start_date": "2012-10-01", "end_date": "2013-09-30"},
						{"start_date": "2013-10-01", "end_date": "2014-09-30"},
						{"start_date": "2014-10-01", "end_date": "2015-09-30"},
						{"start_date": "2015-10-01", "end_date": "2016-09-30"},
						{"start_date": "2016-10-01", "end_date": "2017-09-30"},
						{"start_date": "2017-10-01", "end_date": "2018-09-30"},
						{"start_date": "2007-10-01", "end_date": "2008-09-30"},
						{"start_date": "2008-10-01", "end_date": "2009-09-30"},
						{"start_date": "2009-10-01", "end_date": "2010-09-30"},
						{"start_date": "2010-10-01", "end_date": "2011-09-30"},
						{"start_date": "2011-10-01", "end_date": "2012-09-30"}]
	},
	"limit":60,
	"subawards":true
}

Before Change: 1562 ms
After Change: 1492 ms


/v2/spending_over_time/
REQUEST:
{
	"fields":["Sub-Award ID", "Sub-Award Date", "Sub-Award Amount", "Awarding Agency"],
	"group":"quarter",
	"filters":{
		"keyword":"pizza",
		"place_of_performance_scope": "foreign"
	},
	"limit":60,
	"subawards":true
}

Before change: 978 ms
After change: 960 ms


Download time for Search criteria from ticket (DOD, 2009, Foreign, >5000000)
Before change: 3.58 s
After change: 3.64 s