@tony-sappe 2018-09-13
@IsaacRay Can you include more response times for a performance comparison? There are many permutations of the search filter which cause different models to be used.

Including `or` in the predicate often negatively impacts performance.

Can you also include before and after API and download results? Either listed in this PR or in the Jira ticket?